- Changed the parser to return non-trivial datatypes for uint8, uint16, and uint32.

+++ java/src/tv/porst/swfretools/parser/structures/AssetParser.java
@@ -1,11 +1,12 @@
 package tv.porst.swfretools.parser.structures;
 
 import tv.porst.splib.io.BinaryParser;
+import tv.porst.splib.io.UINT16;
 
 public class AssetParser {
 
 	public static AssetTag parse(final BinaryParser parser) {
-		final int tag = parser.readUInt16();
+		final UINT16 tag = parser.readUInt16();
 		final String name = parser.readString();
 
 		return new AssetTag(tag, name);
