+++ java/src/tv/porst/swfretools/parser/structures/MorphFillStyle.java
@@ -1,10 +1,13 @@
 package tv.porst.swfretools.parser.structures;
 
+import tv.porst.splib.io.UINT16;
+import tv.porst.splib.io.UINT8;
+
 public class MorphFillStyle {
 
-	public MorphFillStyle(final int fillStyleType, final RGB startColor, final RGB endColor,
+	public MorphFillStyle(final UINT8 fillStyleType, final RGB startColor, final RGB endColor,
 			final Matrix startGradientMatrix, final Matrix endGradientMatrix,
-			final MorphGradient gradient, final int bitmapId, final Matrix startBitmapMatrix,
+			final MorphGradient gradient, final UINT16 bitmapId, final Matrix startBitmapMatrix,
 			final Matrix endBitmapMatrix) {
 	}
