+++ java/src/tv/porst/swfretools/parser/tags/DefineBitsJPEG2Parser.java
@@ -2,12 +2,13 @@ package tv.porst.swfretools.parser.tags;
 
 import tv.porst.splib.io.BinaryParser;
 import tv.porst.splib.io.BinaryParserHelpers;
+import tv.porst.splib.io.UINT16;
 import tv.porst.swfretools.parser.structures.RecordHeader;
 
 public class DefineBitsJPEG2Parser {
 
 	public static Tag parse(final RecordHeader header, final BinaryParser parser) {
-		final int characterId = parser.readUInt16();
+		final UINT16 characterId = parser.readUInt16();
 		final byte[] imageData = BinaryParserHelpers.readByteArray(parser, header.getNormalizedLength());
 
 		return new DefineBitsJPEG2Tag(header, characterId, imageData);
