+++ java/src/tv/porst/swfretools/parser/tags/DefineBitsJPEG2Tag.java
@@ -1,10 +1,11 @@
 package tv.porst.swfretools.parser.tags;
 
+import tv.porst.splib.io.UINT16;
 import tv.porst.swfretools.parser.structures.RecordHeader;
 
 public class DefineBitsJPEG2Tag extends Tag {
 
-	public DefineBitsJPEG2Tag(final RecordHeader header, final int characterId,
+	public DefineBitsJPEG2Tag(final RecordHeader header, final UINT16 characterId,
 			final byte[] imageData) {
 		super(header);
 	}
