+++ java/src/tv/porst/swfretools/parser/tags/DefineBitsJPEG3Tag.java
@@ -1,10 +1,11 @@
 package tv.porst.swfretools.parser.tags;
 
+import tv.porst.splib.io.UINT16;
 import tv.porst.swfretools.parser.structures.RecordHeader;
 
 public class DefineBitsJPEG3Tag extends Tag {
 
-	public DefineBitsJPEG3Tag(final RecordHeader header, final int characterId,
+	public DefineBitsJPEG3Tag(final RecordHeader header, final UINT16 characterId,
 			final byte[] jpegData, final byte[] bitmapAlphaData) {
 		super(header);
 	}
