+++ java/src/tv/porst/swfretools/parser/tags/DefineButtonSoundTag.java
@@ -1,15 +1,16 @@
 package tv.porst.swfretools.parser.tags;
 
+import tv.porst.splib.io.UINT16;
 import tv.porst.swfretools.parser.structures.RecordHeader;
 import tv.porst.swfretools.parser.structures.SoundInfo;
 
 public class DefineButtonSoundTag extends Tag {
 
-	public DefineButtonSoundTag(final RecordHeader header, final int buttonID,
-			final int buttonSoundChar0, final SoundInfo buttonSoundInfo0,
-			final int buttonSoundChar1, final SoundInfo buttonSoundInfo1,
-			final int buttonSoundChar2, final SoundInfo buttonSoundInfo2,
-			final int buttonSoundChar3, final SoundInfo buttonSoundInfo3) {
+	public DefineButtonSoundTag(final RecordHeader header, final UINT16 buttonID,
+			final UINT16 buttonSoundChar0, final SoundInfo buttonSoundInfo0,
+			final UINT16 buttonSoundChar1, final SoundInfo buttonSoundInfo1,
+			final UINT16 buttonSoundChar2, final SoundInfo buttonSoundInfo2,
+			final UINT16 buttonSoundChar3, final SoundInfo buttonSoundInfo3) {
 		super(header);
 	}
