+++ java/src/tv/porst/swfretools/parser/tags/DefineFontAlignmentZonesTag.java
@@ -2,12 +2,13 @@ package tv.porst.swfretools.parser.tags;
 
 import java.util.List;
 
+import tv.porst.splib.io.UINT16;
 import tv.porst.swfretools.parser.structures.RecordHeader;
 import tv.porst.swfretools.parser.structures.ZoneRecord;
 
 public class DefineFontAlignmentZonesTag extends Tag {
 
-	public DefineFontAlignmentZonesTag(final RecordHeader header, final int fontID,
+	public DefineFontAlignmentZonesTag(final RecordHeader header, final UINT16 fontID,
 			final int csmTableHint, final int reserved, final List<ZoneRecord> zoneTable) {
 		super(header);
 	}
